Renames breakpoint-match file to match its function name
