+++ lib/breakpointMatch.test.js
@@ -1,4 +1,4 @@
-const breakpointMatch = require('./breakpoint-match');
+const breakpointMatch = require('./breakpointMatch');
 
 describe('Find a breakpoint match: px', () => {
   // Parsed options.
