+++ lib/parse-options.js
@@ -1,5 +1,5 @@
 const { normalizeOptions, varPattern } = require('./normalize-options');
-const breakpointMatch = require('./breakpoint-match');
+const breakpointMatch = require('./breakpointMatch');
 const camelCaseString = require('./utils/camelCaseString');
 
 /**
